+++ packages/lit-html/src/test/directives/style-map_test.ts
@@ -146,7 +146,7 @@ suite('styleMap', () => {
     });
   });
 
-  test('throws when used in NodePart', () => {
+  test('throws when used in ChildPart', () => {
     assert.throws(() => {
       render(html`<div>${styleMap({})}</div>`, container);
     });