+++ packages/lit-ssr/src/test/integration/tests/basic.ts
@@ -57,10 +57,10 @@ export const tests: {[name: string]: SSRTest} = {
   // TODO: add suites (for now, delineating with comments)
 
   /******************************************************
-   * NodePart tests
+   * ChildPart tests
    ******************************************************/
 
-  'NodePart accepts a string': {
+  'ChildPart accepts a string': {
     render(x: unknown) {
       return html`<div>${x}</div>`;
     },
@@ -77,7 +77,7 @@ export const tests: {[name: string]: SSRTest} = {
     stableSelectors: ['div'],
   },
 
-  'NodePart accepts a number': {
+  'ChildPart accepts a number': {
     render(x: unknown) {
       return html`<div>${x}</div>`;
     },
@@ -94,7 +94,7 @@ export const tests: {[name: string]: SSRTest} = {
     stableSelectors: ['div'],
   },
 
-  'NodePart accepts undefined': {
+  'ChildPart accepts undefined': {
     render(x: unknown) {
       return html`<div>${x}</div>`;
     },
@@ -111,7 +111,7 @@ export const tests: {[name: string]: SSRTest} = {
     stableSelectors: ['div'],
   },
 
-  'NodePart accepts null': {
+  'ChildPart accepts null': {
     render(x: unknown) {
       return html`<div>${x}</div>`;
     },
@@ -128,7 +128,7 @@ export const tests: {[name: string]: SSRTest} = {
     stableSelectors: ['div'],
   },
 
-  'NodePart accepts noChange': {
+  'ChildPart accepts noChange': {
     render(x: unknown) {
       return html`<div>${x}</div>`;
     },
@@ -145,7 +145,7 @@ export const tests: {[name: string]: SSRTest} = {
     stableSelectors: ['div'],
   },
 
-  'NodePart accepts nothing': {
+  'ChildPart accepts nothing': {
     render(x: unknown) {
       return html`<div>${x}</div>`;
     },
@@ -162,7 +162,7 @@ export const tests: {[name: string]: SSRTest} = {
     stableSelectors: ['div'],
   },
 
-  'NodePart accepts an object': {
+  'ChildPart accepts an object': {
     render(x: unknown) {
       return html`<div>${x}</div>`;
     },
@@ -181,7 +181,7 @@ export const tests: {[name: string]: SSRTest} = {
     stableSelectors: ['div'],
   },
 
-  'NodePart accepts an object with a toString method': {
+  'ChildPart accepts an object with a toString method': {
     render(x: unknown) {
       return html`<div>${x}</div>`;
     },
@@ -212,7 +212,7 @@ export const tests: {[name: string]: SSRTest} = {
     stableSelectors: ['div'],
   },
 
-  'NodePart accepts a function': {
+  'ChildPart accepts a function': {
     render(x: unknown) {
       return html`<div>${x}</div>`;
     },
@@ -233,7 +233,7 @@ export const tests: {[name: string]: SSRTest} = {
     stableSelectors: ['div'],
   },
 
-  'NodePart accepts TemplateResult': {
+  'ChildPart accepts TemplateResult': {
     render(x: unknown) {
       return html`<div>${html`<span>${x}</span>`}</div>`;
     },
@@ -250,7 +250,7 @@ export const tests: {[name: string]: SSRTest} = {
     stableSelectors: ['div', 'span'],
   },
 
-  'multiple NodeParts, adjacent primitive values': {
+  'multiple ChildParts, adjacent primitive values': {
     render(x: unknown, y: unknown) {
       return html`<div>${x}${y}</div>`;
     },
@@ -267,7 +267,7 @@ export const tests: {[name: string]: SSRTest} = {
     stableSelectors: ['div'],
   },
 
-  'multiple NodeParts, adjacent primitive & TemplateResult': {
+  'multiple ChildParts, adjacent primitive & TemplateResult': {
     render(x: unknown, y: unknown) {
       return html`<div>${x}${html`<span>${y}</span>`}</div>`;
     },
@@ -284,7 +284,7 @@ export const tests: {[name: string]: SSRTest} = {
     stableSelectors: ['div', 'span'],
   },
 
-  'multiple NodeParts, adjacent TemplateResults': {
+  'multiple ChildParts, adjacent TemplateResults': {
     render(x: unknown, y: unknown) {
       return html`<div>
         ${html`<span>${x}</span>`}${html`<span>${y}</span>`}
@@ -303,7 +303,7 @@ export const tests: {[name: string]: SSRTest} = {
     stableSelectors: ['div', 'span'],
   },
 
-  'multiple NodeParts with whitespace': {
+  'multiple ChildParts with whitespace': {
     render(x: unknown, y: unknown) {
       return html`<div>${x} ${y}</div>`;
     },
@@ -334,7 +334,7 @@ export const tests: {[name: string]: SSRTest} = {
     stableSelectors: ['div'],
   },
 
-  'NodePart with trailing whitespace': {
+  'ChildPart with trailing whitespace': {
     render(x: unknown) {
       // prettier-ignore
       return html`<div>${x} </div>`;
@@ -366,7 +366,7 @@ export const tests: {[name: string]: SSRTest} = {
     stableSelectors: ['div'],
   },
 
-  'NodePart accepts array with strings': {
+  'ChildPart accepts array with strings': {
     render(words: string[]) {
       return html`<div>${words}</div>`;
     },
@@ -383,7 +383,7 @@ export const tests: {[name: string]: SSRTest} = {
     stableSelectors: ['div'],
   },
 
-  'NodePart accepts array with strings, updated with fewer items': {
+  'ChildPart accepts array with strings, updated with fewer items': {
     render(words: string[]) {
       return html`<div>${words}</div>`;
     },
@@ -401,7 +401,7 @@ export const tests: {[name: string]: SSRTest} = {
     stableSelectors: ['div'],
   },
 
-  'NodePart accepts array with strings, updated with more items': {
+  'ChildPart accepts array with strings, updated with more items': {
     render(words: string[]) {
       return html`<div>${words}</div>`;
     },
@@ -419,7 +419,7 @@ export const tests: {[name: string]: SSRTest} = {
     stableSelectors: ['div'],
   },
 
-  'NodePart accepts array with templates': {
+  'ChildPart accepts array with templates': {
     render(words: string[]) {
       return html`<ol>
         ${words.map((w) => html`<li>${w}</li>`)}
@@ -438,7 +438,7 @@ export const tests: {[name: string]: SSRTest} = {
     stableSelectors: ['ol', 'li'],
   },
 
-  'NodePart accepts simple directive': () => {
+  'ChildPart accepts simple directive': () => {
     const basic = directive(
       class extends Directive {
         count = 0;
@@ -470,7 +470,7 @@ export const tests: {[name: string]: SSRTest} = {
     };
   },
 
-  'NodePart accepts nested directives': () => {
+  'ChildPart accepts nested directives': () => {
     const aDirective = directive(
       class extends Directive {
         render(bool: boolean, v: unknown) {
@@ -517,7 +517,7 @@ export const tests: {[name: string]: SSRTest} = {
     };
   },
 
-  'NodePart accepts directive: repeat (with strings)': {
+  'ChildPart accepts directive: repeat (with strings)': {
     render(words: string[]) {
       return html`${repeat(words, (word, i) => `(${i} ${word})`)}`;
     },
@@ -534,7 +534,7 @@ export const tests: {[name: string]: SSRTest} = {
     stableSelectors: [],
   },
 
-  'NodePart accepts directive: repeat (with templates)': {
+  'ChildPart accepts directive: repeat (with templates)': {
     render(words: string[]) {
       return html`${repeat(words, (word, i) => html`<p>${i}) ${word}</p>`)}`;
     },
@@ -551,7 +551,7 @@ export const tests: {[name: string]: SSRTest} = {
     stableSelectors: ['p'],
   },
 
-  'NodePart accepts directive: cache': {
+  'ChildPart accepts directive: cache': {
     render(bool: boolean) {
       return html`${cache(bool ? html`<p>true</p>` : html`<b>false</b>`)}`;
     },
@@ -572,7 +572,7 @@ export const tests: {[name: string]: SSRTest} = {
     stableSelectors: [],
   },
 
-  'NodePart accepts directive: guard': () => {
+  'ChildPart accepts directive: guard': () => {
     let guardedCallCount = 0;
     const guardedTemplate = (bool: boolean) => {
       guardedCallCount++;
@@ -609,7 +609,7 @@ export const tests: {[name: string]: SSRTest} = {
     };
   },
 
-  'NodePart accepts directive: until (primitive)': {
+  'ChildPart accepts directive: until (primitive)': {
     render(...args) {
       return html`<div>${until(...args)}</div>`;
     },
@@ -626,7 +626,7 @@ export const tests: {[name: string]: SSRTest} = {
     stableSelectors: ['div'],
   },
 
-  'NodePart accepts directive: until (promise, primitive)': () => {
+  'ChildPart accepts directive: until (promise, primitive)': () => {
     let resolve: (v: string) => void;
     const promise = new Promise((r) => (resolve = r));
     return {
@@ -651,7 +651,7 @@ export const tests: {[name: string]: SSRTest} = {
     };
   },
 
-  'NodePart accepts directive: until (promise, promise)': () => {
+  'ChildPart accepts directive: until (promise, promise)': () => {
     let resolve1: (v: string) => void;
     let resolve2: (v: string) => void;
     const promise1 = new Promise((r) => (resolve1 = r));
@@ -688,7 +688,7 @@ export const tests: {[name: string]: SSRTest} = {
 
   // TODO(kschaaf): Enable once async directives are implemented
 
-  // 'NodePart accepts directive: asyncAppend': () => {
+  // 'ChildPart accepts directive: asyncAppend': () => {
   //   const iterable = new TestAsyncIterable();
   //   return {
   //     render(iterable) {
@@ -718,7 +718,7 @@ export const tests: {[name: string]: SSRTest} = {
   //   };
   // },
 
-  // 'NodePart accepts directive: asyncReplace': () => {
+  // 'ChildPart accepts directive: asyncReplace': () => {
   //   const iterable = new TestAsyncIterable();
   //   return {
   //     render(iterable) {
@@ -748,7 +748,7 @@ export const tests: {[name: string]: SSRTest} = {
   //   };
   // },
 
-  'NodePart accepts directive: ifDefined (undefined)': {
+  'ChildPart accepts directive: ifDefined (undefined)': {
     render(v) {
       return html`<div>${ifDefined(v)}</div>`;
     },
@@ -765,7 +765,7 @@ export const tests: {[name: string]: SSRTest} = {
     stableSelectors: ['div'],
   },
 
-  'NodePart accepts directive: ifDefined (defined)': {
+  'ChildPart accepts directive: ifDefined (defined)': {
     render(v) {
       return html`<div>${ifDefined(v)}</div>`;
     },
@@ -782,7 +782,7 @@ export const tests: {[name: string]: SSRTest} = {
     stableSelectors: ['div'],
   },
 
-  'NodePart accepts directive: unsafeHTML': {
+  'ChildPart accepts directive: unsafeHTML': {
     render(v) {
       return html`<div>${unsafeHTML(v)}</div>`;
     },
@@ -799,7 +799,7 @@ export const tests: {[name: string]: SSRTest} = {
     stableSelectors: ['div'],
   },
 
-  'NodePart accepts directive: unsafeSVG': {
+  'ChildPart accepts directive: unsafeSVG': {
     render(v) {
       return html`<svg>${unsafeSVG(v)}</svg>`;
     },
@@ -3284,7 +3284,7 @@ export const tests: {[name: string]: SSRTest} = {
    * Mixed part tests
    ******************************************************/
 
-  'NodeParts & AttributeParts on adjacent nodes': {
+  'ChildParts & AttributeParts on adjacent nodes': {
     render(x, y) {
       return html`<div attr="${x}">${x}</div>
         <div attr="${y}">${y}</div>`;
@@ -3302,7 +3302,7 @@ export const tests: {[name: string]: SSRTest} = {
     stableSelectors: ['div'],
   },
 
-  'NodeParts & AttributeParts on nested nodes': {
+  'ChildParts & AttributeParts on nested nodes': {
     render(x, y) {
       return html`<div attr="${x}">
         ${x}
@@ -3322,7 +3322,7 @@ export const tests: {[name: string]: SSRTest} = {
     stableSelectors: ['div'],
   },
 
-  'NodeParts & AttributeParts soup': {
+  'ChildParts & AttributeParts soup': {
     render(x, y, z) {
       return html`text:${x}
         <div>${x}</div>